stt-worker/requirements.txt
# STT Worker Dependencies
faster-whisper>=1.0.0
firebase-admin>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydub>=0.25.1
numpy>=1.24.0

# Auto-processor (optional, for automatic processing)
# No additional dependencies needed
